write c program for loading generated points

+++ pointplot.py
@@ -4,7 +4,7 @@ points = []
 with open("points.txt", "r") as f:
     lines = f.readlines()[1:]
     for line in lines:
-        x, y = [float(i) for i in line.split()]
+        x, y = [int(i) for i in line.split()]
         points.append((x, y))
 
 # Separate the points into x and y coordinates
